Legg dialogporten dataklasser i egen mappe
