--- src/main/kotlin/no/nav/helsearbeidsgiver/dialogporten/CreateDialogRequest.kt
@@ -1,223 +1,0 @@
-package no.nav.helsearbeidsgiver.dialogporten
-
-import kotlinx.serialization.Serializable
-import java.util.UUID
-
-@Serializable
-data class CreateDialogRequest(
-    val serviceResource: String,
-    val party: String,
-    val externalRefererence: String,
-    val status: String,
-    val content: Content,
-    val guiActions: List<GuiAction>,
-    val transmissions: List<Transmission>,
-)
-
-@Serializable
-data class AddTransmissionsRequest(
-    val transmissions: List<Transmission>,
-) {
-    val op = "add"
-    val path = "/transmissions"
-}
-
-@Serializable
-data class GuiAction(
-    val action: String,
-    val url: String,
-    val title: List<ContentValueItem>,
-) {
-    @Suppress("unused")
-    val priority = "Primary"
-
-    @Suppress("unused")
-    val isDeleteDialogAction = false
-}
-
-@Serializable
-data class Content(
-    val title: ContentValue,
-    val summary: ContentValue,
-)
-
-@Serializable
-data class ContentValue(
-    val value: List<ContentValueItem>,
-) {
-    @Suppress("unused")
-    val mediaType: String = "text/plain"
-}
-
-@Serializable
-data class ContentValueItem(
-    val value: String,
-) {
-    @Suppress("unused")
-    val languageCode: String = "nb"
-}
-
-@Serializable
-data class Transmission(
-    val type: TransmissionType,
-    val sender: Sender,
-    val content: Content,
-    val attachments: List<Attachment>,
-) {
-    @Serializable
-    data class Sender(
-        val actorType: String,
-    )
-
-    @Serializable
-    enum class TransmissionType {
-        // For general information, not related to any submissions
-        Information,
-
-        // Feedback/receipt accepting a previous submission
-        Acceptance,
-
-        // Feedback/error message rejecting a previous submission
-        Rejection,
-
-        // Question/request for more information
-        Request,
-    }
-
-    @Serializable
-    data class Attachment(
-        val displayName: List<ContentValueItem>,
-        val urls: List<Url>,
-    )
-
-    @Serializable
-    data class Url(
-        val url: String,
-        val mediaType: String,
-        val consumerType: AttachmentUrlConsumerType,
-    )
-
-    @Serializable
-    enum class AttachmentUrlConsumerType {
-        Gui,
-        Api,
-    }
-}
-
-fun lagContentValue(verdi: String) =
-    ContentValue(
-        value =
-            listOf(
-                ContentValueItem(
-                    value = verdi,
-                ),
-            ),
-    )
-
-fun lagGuiAction(
-    url: String,
-    tittel: String,
-) = GuiAction(
-    action = "read",
-    url = url,
-    title =
-        listOf(
-            ContentValueItem(
-                value = tittel,
-            ),
-        ),
-)
-
-fun lagCreateDialogRequest(
-    ressurs: String,
-    orgnr: String,
-    status: String,
-    tittel: String,
-    sammendrag: String,
-    url: String,
-    knappTittel: String,
-): CreateDialogRequest =
-    CreateDialogRequest(
-        serviceResource = "urn:altinn:resource:$ressurs",
-        party = "urn:altinn:organization:identifier-no:$orgnr",
-        status = status,
-        externalRefererence = UUID.randomUUID().toString(),
-        content = Content(lagContentValue(tittel), lagContentValue(sammendrag)),
-        guiActions = listOf(lagGuiAction(url, knappTittel)),
-        transmissions = emptyList(),
-    )
-
-fun lagNyDialogMedSykmeldingRequest(
-    ressurs: String,
-    orgnr: String,
-    dialogTittel: String,
-    dialogSammendrag: String,
-    sykmeldingId: UUID,
-    sykmeldingJsonUrl: String,
-): CreateDialogRequest =
-    CreateDialogRequest(
-        serviceResource = "urn:altinn:resource:$ressurs",
-        party = "urn:altinn:organization:identifier-no:$orgnr",
-        status = "New",
-        externalRefererence = sykmeldingId.toString(),
-        content = Content(lagContentValue(dialogTittel), lagContentValue(dialogSammendrag)),
-        guiActions = emptyList(),
-        transmissions =
-            listOf(
-                lagVedleggTransmission(
-                    transmissionTittel = "Sykmelding",
-                    transmissionSammendrag = "Sykmelding",
-                    vedleggNavn = "Sykmelding.json",
-                    vedleggUrl = sykmeldingJsonUrl,
-                    vedleggMediaType = "application/json",
-                    vedleggConsumerType = Transmission.AttachmentUrlConsumerType.Api,
-                ),
-            ),
-    )
-
-fun oppdaterDialogMedSoknadRequest(soknadJsonUrl: String): AddTransmissionsRequest =
-    AddTransmissionsRequest(
-        transmissions =
-            listOf(
-                lagVedleggTransmission(
-                    transmissionTittel = "Søknad",
-                    transmissionSammendrag = "Søknad",
-                    vedleggNavn = "Søknad.json",
-                    vedleggUrl = soknadJsonUrl,
-                    vedleggMediaType = "application/json",
-                    vedleggConsumerType = Transmission.AttachmentUrlConsumerType.Api,
-                ),
-            ),
-    )
-
-private fun lagVedleggTransmission(
-    transmissionTittel: String,
-    transmissionSammendrag: String,
-    vedleggNavn: String,
-    vedleggUrl: String,
-    vedleggMediaType: String,
-    vedleggConsumerType: Transmission.AttachmentUrlConsumerType,
-): Transmission =
-    Transmission(
-        type = Transmission.TransmissionType.Information,
-        sender = Transmission.Sender("ServiceOwner"),
-        content =
-            Content(
-                title = lagContentValue(transmissionTittel),
-                summary = lagContentValue(transmissionSammendrag),
-            ),
-        attachments =
-            listOf(
-                Transmission.Attachment(
-                    displayName = listOf(ContentValueItem(vedleggNavn)),
-                    urls =
-                        listOf(
-                            Transmission.Url(
-                                url = vedleggUrl,
-                                mediaType = vedleggMediaType,
-                                consumerType = vedleggConsumerType,
-                            ),
-                        ),
-                ),
-            ),
-    )
