+++ src/main/kotlin/no/nav/helsearbeidsgiver/dialogporten/domene/AddTransmissionsRequest.kt
@@ -1,0 +1,12 @@
+package no.nav.helsearbeidsgiver.dialogporten.domene
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AddTransmissionsRequest(
+    val transmissions: List<Transmission>,
+) {
+    @Suppress("unused")
+    val op = "add"
+    val path = "/transmissions"
+}
