+++ src/main/kotlin/no/nav/helsearbeidsgiver/dialogporten/domene/CreateDialogRequest.kt
@@ -1,0 +1,14 @@
+package no.nav.helsearbeidsgiver.dialogporten.domene
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CreateDialogRequest(
+    val serviceResource: String,
+    val party: String,
+    val externalRefererence: String,
+    val status: String,
+    val content: Content,
+    val guiActions: List<GuiAction>,
+    val transmissions: List<Transmission>,
+)
