--- src/test/resources/create-dialog-response/created.json
@@ -1,1 +1,0 @@
-"0194cb3a-6f4e-7707-a506-a1db2b5c37fa"
